requirements.txt
slack-sdk>=3.19.0
python-dotenv>=0.19.0
